backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
pytest==8.3.5
behave==1.2.6
pytest-cov==4.1.0
httpx==0.26.0
python-dotenv==1.0.0
alembic==1.13.1
